Pinocchio file changes to run maven inside docker
Summary:
Frontend plugin required maven 3.1.0 or later, so explicitly installing the maven 3.3.9.
Also, the docker container couldn't explictly find the .jks file hence had to create it explicitly as a pre-build command.
Support for using config file based on the Uber runtime environment and added the ability to run uworc on the specified Uber http port.
Added a health check resource that can be later extended to check Piper and Mysql database health.

Reviewers: #uworc, akira

Reviewed By: #uworc, akira

Subscribers: akira, jenkins

Differential Revision: https://code.uberinternal.com/D2154123