Optimize and improve the shell scripts
Summary:
Removed the command line arguments to the various shell scripts that was passing the uber-runtime-environment & uber-port-http explicitly.
Setting the default environment as 'dev' and the default port as '8080'.
The shell scripts are now what they were in the open source repo.
Starting the streamline server in foreground using a new command line option - start-fg.

Reviewers: #uworc, harshach

Reviewed By: #uworc, harshach

Subscribers: harshach

Differential Revision: https://code.uberinternal.com/D2236559